backend queries turned to strings

+++ controllers/searchController.js
@@ -48,8 +48,8 @@ exports.searchBooks = catchAsync(async (req, res, next) => {
     books.date_created,
     books.content_rating,
     workGenres.genre_array,
-    (SELECT COUNT(*) FROM issues WHERE issues.book_id = books.id) AS total_issues,
-    (SELECT COUNT(*) FROM issue_assets WHERE issue_assets.book_id = books.id) AS total_book_pages
+    (SELECT COUNT(*) FROM issues WHERE issues.book_id = books.id)::integer AS total_issues,
+    (SELECT COUNT(*) FROM issue_assets WHERE issue_assets.book_id = books.id)::integer AS total_book_pages
    `,
     searchBook.query,
     searchBook.parameterizedValues,
@@ -103,8 +103,8 @@ exports.searchIssues = catchAsync(async (req, res) => {
     books.url_slug,
     books.content_rating,
     workGenres.genre_array,
-    (SELECT COUNT(*) FROM issues WHERE issues.book_id = books.id) AS total_issues,
-    (SELECT COUNT(*) FROM issue_assets WHERE issue_assets.issue_id = issues.id) AS total_issue_pages
+    (SELECT COUNT(*) FROM issues WHERE issues.book_id = books.id)::integer AS total_issues,
+    (SELECT COUNT(*) FROM issue_assets WHERE issue_assets.issue_id = issues.id)::integer AS total_issue_pages
     `,
     searchIssue.query,
     searchIssue.parameterizedValues,
